props/agents: cleanup — drop AF-vestigial indirection, rename handler→middleware
- rename notification_handler.py → notification_middleware.py (it's a chat middleware now)
- GraderAbortError: BaseException → Exception (BaseException was only to bypass agent_core's
  loop; it's now raised outside the agent)
- remove dead _make_gt_ref; remove unused run_until_done(**run_options)
- make_agent tools: list[FunctionTool]; document direct_tool's schema-supplied FunctionTool mode
